9OS 12 - Alpha 4 (READ DESCRIPTION)
This release REQUIRES a U:\ to use the user functionality. Create a 64 MB VHD, and mount it as U:\. Login tables, configuration files & other user data will be stored here. I would recommend at least 128 MB. It is not required, but to use all of the new features introduced, you will need the drive.